pydanticai_mcp/pyproject.toml
[project]
name = "pydanticai-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "fastmcp>=2.3.3",
    "pydantic-ai>=0.1.11",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.4",
]
